# swaggerstrayV3.json
{
	"openapi": "3.0.0",
	"info": {
	  "title": "Stray",
	  "contact": {},
	  "version": "1.0"
	},
	"servers": [
	  {
		"url": "http://localhost:3000/api/v1",
		"variables": {}
	  }
	],
	"paths": {
	  "/admin/orders": {
		"post": {
		  "tags": [
			"Order"
		  ],
		  "summary": "createOrder",
		  "description": "createOrder",
		  "operationId": "createOrder",
		  "parameters": [
			{
			  "name": "Authorization",
			  "in": "header",
			  "description": "",
			  "style": "simple",
			  "schema": {
				"type": "string",
				"default": "Bearer {token}"
			  }
			}
		  ],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "allOf": [
					{
					  "$ref": "#/components/schemas/createOrderRequest"
					},
					{
					  "example": {
						"promo": "MDK",
						"sender_addres": "fdsafasdfadsfasdM",
						"receiver_addres": "ACEsdfasdfasdfJ",
						"customer_name": "Akhmad",
						"items": [
						  {
							"id_item": "1",
							"item_quantity": "5"
						  }
						]
					  }
					}
				  ]
				},
				"example": {
				  "promo": "MDK",
				  "sender_addres": "fdsafasdfadsfasdM",
				  "receiver_addres": "ACEsdfasdfasdfJ",
				  "customer_name": "Akhmad",
				  "items": [
					{
					  "id_item": "1",
					  "item_quantity": "5"
					}
				  ]
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {},
			  "content": {}
			}
		  },
		  "deprecated": false
		},
		"get": {
		  "tags": [
			"Order"
		  ],
		  "summary": "getAllOrder",
		  "description": "getAllOrder",
		  "operationId": "getAllOrder",
		  "parameters": [
			{
			  "name": "Authorization",
			  "in": "header",
			  "description": "",
			  "style": "simple",
			  "schema": {
				"type": "string",
				"default": "Bearer {token}"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {},
			  "content": {}
			}
		  },
		  "deprecated": false
		}
	  },
...
	  "/member/orders/34": {
		"patch": {
		  "tags": [
			"Order"
		  ],
		  "summary": "updateOrder",
		  "description": "updateOrder",
		  "operationId": "updateOrder",
		  "parameters": [
			{
			  "name": "Authorization",
			  "in": "header",
			  "description": "",
			  "style": "simple",
			  "schema": {
				"type": "string",
				"default": "Bearer {token}"
			  }
			}
		  ],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "allOf": [
					{
					  "$ref": "#/components/schemas/updateOrderRequest"
					},
					{
					  "example": {
						"sender_addres": "Makassar ji abca",
						"status": "Paid"
					  }
					}
				  ]
				},
				"example": {
				  "sender_addres": "Makassar ji abca",
				  "status": "Paid"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {},
			  "content": {}
			}
		  },
		  "deprecated": false
		}
	  },
...
	  "/member/users": {
		"get": {
		  "tags": [
			"User"
		  ],
		  "summary": "userById",
		  "description": "userById",
		  "operationId": "userById",
		  "parameters": [
			{
			  "name": "Authorization",
			  "in": "header",
			  "description": "",
			  "style": "simple",
			  "schema": {
				"type": "string",
				"default": "Bearer {token}"
			  }
			}
		  ],
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {},
			  "content": {}
			}
		  },
		  "deprecated": false
		},
		"patch": {
		  "tags": [
			"User"
		  ],
		  "summary": "updateUser",
		  "description": "updateUser",
		  "operationId": "updateUser",
		  "parameters": [
			{
			  "name": "Authorization",
			  "in": "header",
			  "description": "",
			  "style": "simple",
			  "schema": {
				"type": "string",
				"default": "Bearer {token}"
			  }
			}
		  ],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "allOf": [
					{
					  "$ref": "#/components/schemas/updateUserRequest"
					},
					{
					  "example": {
						"name": "Akhmad Maknur lagi"
					  }
					}
				  ]
				},
				"example": {
				  "name": "Akhmad Maknur lagi"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {},
			  "content": {}
			}
		  },
		  "deprecated": false
		}
	  },
...
	  "/auth/refresh_token": {
		"post": {
		  "tags": [
			"Misc"
		  ],
		  "summary": "RefreshToken",
		  "description": "Refresh Token",
		  "operationId": "RefreshToken",
		  "parameters": [
			{
			  "name": "Authorization",
			  "in": "header",
			  "description": "",
			  "style": "simple",
			  "schema": {
				"type": "string",
				"default": "Bearer {token}"
			  }
			}
		  ],
		  "requestBody": {
			"description": "",
			"content": {
			  "application/json": {
				"schema": {
				  "allOf": [
					{
					  "$ref": "#/components/schemas/RefreshTokenRequest"
					},
					{
					  "example": {
						"email": "akhmad.maknur@gmail.com",
						"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFraG1hZC5tYWtudXJAZ21haWwxLmNvbSIsInVzZXJfaWQiOjIsImlhdCI6MTY2MDg4Nzk1MiwiZXhwIjoxNjYwODkxNTUyfQ.EvFi8la6yv84YsENWRZQPvJMMA8zfrseGj7B8FLuD0I"
					  }
					}
				  ]
				},
				"example": {
				  "email": "akhmad.maknur@gmail.com",
				  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFraG1hZC5tYWtudXJAZ21haWwxLmNvbSIsInVzZXJfaWQiOjIsImlhdCI6MTY2MDg4Nzk1MiwiZXhwIjoxNjYwODkxNTUyfQ.EvFi8la6yv84YsENWRZQPvJMMA8zfrseGj7B8FLuD0I"
				}
			  }
			},
			"required": true
		  },
		  "responses": {
			"200": {
			  "description": "",
			  "headers": {},
			  "content": {}
			}
		  },
		  "deprecated": false
		}
	  },
...
	},
	"components": {
	  "schemas": {
		"createOrderRequest": {
		  "title": "createOrderRequest",
		  "required": [
			"promo",
			"sender_addres",
			"receiver_addres",
			"customer_name",
			"items"
		  ],
		  "type": "object",
		  "properties": {
			"promo": {
			  "type": "string"
			},
			"sender_addres": {
			  "type": "string"
			},
			"receiver_addres": {
			  "type": "string"
			},
			"customer_name": {
			  "type": "string"
			},
			"items": {
			  "type": "array",
			  "items": {
				"$ref": "#/components/schemas/Item"
			  },
			  "description": ""
			}
		  },
		  "example": {
			"promo": "MDK",
			"sender_addres": "fdsafasdfadsfasdM",
			"receiver_addres": "ACEsdfasdfasdfJ",
			"customer_name": "Akhmad",
			"items": [
			  {
				"id_item": "1",
				"item_quantity": "5"
			  }
			]
		  }
		},
		"Item": {
		  "title": "Item",
		  "required": [
			"id_item",
			"item_quantity"
		  ],
		  "type": "object",
		  "properties": {
			"id_item": {
			  "type": "string"
			},
			"item_quantity": {
			  "type": "string"
			}
		  },
		  "example": {
			"id_item": "1",
			"item_quantity": "5"
		  }
		},
...
		"updateUserRequest": {
		  "title": "updateUserRequest",
		  "required": [
			"name"
		  ],
		  "type": "object",
		  "properties": {
			"name": {
			  "type": "string"
			}
		  },
		  "example": {
			"name": "Akhmad Maknur lagi"
		  }
		},
...
		"updateOrderRequest": {
		  "title": "updateOrderRequest",
		  "required": [
			"sender_addres",
			"status"
		  ],
		  "type": "object",
		  "properties": {
			"sender_addres": {
			  "type": "string"
			},
			"status": {
			  "type": "string"
			}
		  },
		  "example": {
			"sender_addres": "Makassar ji abca",
			"status": "Paid"
		  }
		},
...
		"RefreshTokenRequest": {
		  "title": "RefreshTokenRequest",
		  "required": [
			"email",
			"token"
		  ],
		  "type": "object",
		  "properties": {
			"email": {
			  "type": "string"
			},
			"token": {
			  "type": "string"
			}
		  },
		  "example": {
			"email": "akhmad.maknur@gmail.com",
			"token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFraG1hZC5tYWtudXJAZ21haWwxLmNvbSIsInVzZXJfaWQiOjIsImlhdCI6MTY2MDg4Nzk1MiwiZXhwIjoxNjYwODkxNTUyfQ.EvFi8la6yv84YsENWRZQPvJMMA8zfrseGj7B8FLuD0I"
		  }
		},
...
	  }
	},
...
  }